% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_psi_counts.R
\name{add_psi_counts}
\alias{add_psi_counts}
\title{Add count data to a PSI SummarizedExperiment object}
\usage{
add_psi_counts(se_psi, se_gene)
}
\arguments{
\item{se_psi}{A PSI SummarizedExperiment object returned by the
\code{\link{transcript_to_psi}} function.}

\item{se_gene}{A gene-level SummarizedExperiment object returned by the
\code{\link{tcc_to_gene}} function. It must be compatible with the se_psi
object (i.e. they must originate from the same TCC data).}
}
\value{
A copy of the se_psi PSI SummarizedExperiment object with count
assays added.
}
\description{
Adds two assays ('counts' and 'other_counts') to a PSI SummarizedExperiment
object, making it suitable for downstream analysis using the DEXSeq package.
}
