% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_intron_strand.R
\name{assign_intron_strand}
\alias{assign_intron_strand}
\title{Assigning strand to introns}
\usage{
assign_intron_strand(
  intron_granges,
  anno_data,
  genome_fasta_file,
  min_intron_length = 30,
  known_intron_motifs = c("GT-AG"),
  rescue_annotated_introns = FALSE
)
}
\arguments{
\item{intron_granges}{A GRanges object storing intron positions}

\item{anno_data}{A list containing genome annotation data returned by
the \code{\link{prepare_reference_annotations}} function}

\item{genome_fasta_file}{A string containing a genome FASTA file path}

\item{min_intron_length}{An integer scalar specifying the minimal length
of introns to assign strand to}

\item{known_intron_motifs}{A character vector specifying the known intron
motifs}

\item{rescue_annotated_introns}{A logical scalar specifying if introns
found in genome annotations should be kept even if they don't have known
intron motifs}
}
\value{
A copy of the intron_granges GRanges object with intron strands
assigned
}
\description{
Assign strand to the intron GRanges object based on known intron motifs
}
\keyword{internal}
