% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_to_tcc.R
\name{bam_to_tcc}
\alias{bam_to_tcc}
\title{Prepare a TCC SummarizedExperiment object}
\usage{
bam_to_tcc(
  bam_files,
  transcript_data,
  run_mode = "de_novo_loose",
  min_read_count = 1,
  min_relative_expression = 0,
  extend_spliced_transcripts = 100,
  is_single_cell = FALSE,
  barcode_tag = "BC",
  chunk_size = 1e+06,
  ncpu = 1
)
}
\arguments{
\item{bam_files}{A named character vector containing BAM file paths.}

\item{transcript_data}{A named list containing transcript data returned by
the \code{\link{prepare_transcripts}} function.}

\item{run_mode}{A string specifying the mode for choosing the transcript set
('strict', 'de_novo_strict', 'de_novo_loose' or 'de_novo_full').}

\item{min_read_count}{An integer scalar specifying the read count threshold
for transcripts extracted from the BAM files.}

\item{min_relative_expression}{A numeric scalar specifying the relative
expression threshold for transcripts extracted from the BAM files.}

\item{extend_spliced_transcripts}{An integer scalar specifying the number of
base pairs by which transcript starts and ends are extended for spliced read
compatibility search.}

\item{is_single_cell}{A logical scalar specifying if the BAM files contain
single cell data.}

\item{barcode_tag}{A string specifying the name of the BAM file tag
containing cell barcodes.}

\item{chunk_size}{An integer scalar specifying the chunk size for reading
the BAM files.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A SummarizedExperiment object containing TCC annotation and
quantification data.
}
\description{
Prepares a TCC (Transcript Compatibility Counts) SummarizedExperiment object
for the given BAM files and transcript set.
}
