% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_psi_ratio_matrix.R
\name{calculate_psi_ratio_matrix}
\alias{calculate_psi_ratio_matrix}
\title{Calculate PSI count to mean permuted PSI count ratio matrix}
\usage{
calculate_psi_ratio_matrix(
  se_tcc,
  pseudotime_matrix,
  psi_events,
  window_sizes,
  window_steps,
  trim = 0,
  n_perm = 100,
  ncpu = 1
)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{bam_to_tcc}} function.}

\item{pseudotime_matrix}{A numeric matrix containing the pseudotime values
for each cell (rows) in different trajectories (columns). Cells not
belonging to given trajectory should be denoted using NA values.}

\item{psi_events}{A character vector specifying the PSI events to calculate
the ratios for.}

\item{window_sizes}{A named integer vector specifying the window size for
each trajectory.}

\item{window_steps}{A named integer vector specifying the window step for
each trajectory.}

\item{trim}{A numeric scalar specifying the fraction (0 to 0.5) of cells
to be trimmed from each end of the pseudotime spectrum for each trajectory.}

\item{n_perm}{An integer scalar specifying the number of PSI count
permutations to calculate.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A numeric matrix containing the PSI count to mean permuted PSI count
ratio values.
}
\description{
Calculates PSI count to mean permuted PSI count ratio matrix for pseudotime
window data. This function is designed for preparing data to be visualized
as a heatmap, and might take a long time to run - see the vignettes for an
example.
}
