% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_relative_expression.R
\name{calculate_relative_expression}
\alias{calculate_relative_expression}
\title{Calculate relative expression}
\usage{
calculate_relative_expression(se)
}
\arguments{
\item{se}{A SummarizedExperiment or SingleCellExperiment object containing
a 'tpm' assay and a 'gene_id' column in its row metadata}
}
\value{
A matrix (regular or sparse, depending on the input data) containing
the calculated relative expression values
}
\description{
Calculates relative expression from TPM values stored in a SummarizedExperiment
or SingleCellExperiment object
}
\keyword{internal}
