% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tpm.R
\name{calculate_tpm}
\alias{calculate_tpm}
\title{Calculate TPM values}
\usage{
calculate_tpm(se)
}
\arguments{
\item{se}{A SummarizedExperiment or SingleCellExperiment object containing
a 'counts' assay.}
}
\value{
A matrix (regular or sparse, depending on the input data) containing
the calculated TPM values.
}
\description{
Calculates TPM values from counts stored in a SummarizedExperiment
or SingleCellExperiment object.
}
\keyword{internal}
