% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_splicing_compatibility.R
\name{check_splicing_compatibility}
\alias{check_splicing_compatibility}
\title{Splicing compatibility check}
\usage{
check_splicing_compatibility(subject, target)
}
\arguments{
\item{subject}{A list of vectors storing intron indices for subject
transcripts}

\item{target}{A list of vectors storing intron indices for target
transcripts}
}
\value{
A data frame storing indices of compatible subject and target
transcripts
}
\description{
Check splicing compatibility between subject and target transcripts
}
\keyword{internal}
