% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM}
\alias{EM}
\title{EM algorithm}
\usage{
EM(
  counts,
  compatibility_matrix,
  use_length_normalization,
  tx_effective_lengths,
  maxiter,
  conv
)
}
\arguments{
\item{counts}{A diagonal matrix of TCC read counts}

\item{use_length_normalization}{A logical scalar specifying if normalization
using effective transcript lengths should be used during EM}

\item{tx_effective_lengths}{A numeric vector of effective transcript lengths}

\item{maxiter}{An integer scalar specifying the maximum number of EM iterations}

\item{conv}{A numeric scalar specifying the EM convergence threshold}

\item{compatiblity_matrix}{A compatibility matrix between TCCs and transcripts}
}
\value{
A numeric vector of transcript read counts estimated using EM
}
\description{
Run EM using RcppArmadillo
}
\keyword{internal}
