% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_psi_events.R
\name{filter_psi_events}
\alias{filter_psi_events}
\title{Filter PSI events}
\usage{
filter_psi_events(
  se_psi,
  gene_ids = NULL,
  exclude_tss = TRUE,
  exclude_tes = TRUE,
  min_mean_psi = 0.025,
  max_mean_psi = 0.975,
  min_alt_psi_count = 30,
  min_incl_psi = 0.1,
  min_incl_psi_count = 30
)
}
\arguments{
\item{se_psi}{A PSI SummarizedExperiment object returned by a function from
the \code{\link{Isosceles-package}}.}

\item{gene_ids}{A character vector specifying the gene IDs to restrict
the results to (ignored if set to NULL).}

\item{exclude_tss}{A logical scalar specifying if TSS (transcription start
sites) regions should be excluded from the results.}

\item{exclude_tes}{A logical scalar specifying if TES (transcription end
sites) regions should be excluded from the results.}

\item{min_mean_psi}{A numeric scalar specifying the minimum value threshold
for mean PSI values across all cells.}

\item{max_mean_psi}{A numeric scalar specifying the maximum value threshold
for mean PSI values across all cells.}

\item{min_alt_psi_count}{An integer scalar specifying the required minimum
number of cells with PSI values different than 0, 0.5 or 1.}

\item{min_incl_psi}{A numeric scalar specifying the minimum PSI value for
cells considered to include given PSI event.}

\item{min_incl_psi_count}{An integer scalar specifying the required minimum
number of cells including given PSI event.}
}
\value{
A vector of filtered PSI event identifiers.
}
\description{
Prepares a vector of filtered PSI events suitable to be used with the
\code{\link{psi_to_dexseq}} function.
}
