% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_iso_switch.R
\name{find_iso_switch}
\alias{find_iso_switch}
\title{Find isoform switching events}
\usage{
find_iso_switch(se, cell_labels, min_fdr = 0.05, ncpu = 1)
}
\arguments{
\item{se}{A transcript-level SummarizedExperiment object returned by the
\code{\link{tcc_to_transcript}} function. The object must contain normalized
data stored in the 'logcounts' assay, which can be prepared using functions
from the scuttle package.}

\item{cell_labels}{A vector or a factor containing cell labels acting as a
grouping variable.}

\item{min_fdr}{A numeric scalar specifying the FDR threshold for filtering
the results.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A data frame containing the following columns:
\describe{
\item{transcript_id}{Isosceles transcript ID}
\item{compatible_tx}{comma-separated list of annotated transcript IDs compatible with the Isosceles transcript}
\item{gene_id}{gene ID}
\item{gene_name}{gene symbol}
\item{pvalue}{p-value from the Wilcoxon test performed by the findMarkers function}
\item{fdr}{false discovery rate (FDR) value from the Wilcoxon test performed by the findMarkers function}
\item{auc}{area under the curve (AUC) value from the Wilcoxon test performed by the findMarkers function}
\item{group_1}{label of the cell group in which the transcript is upregulated}
\item{group_2}{label of the cell group compared to which the transcript is upregulated}
\item{contrast}{label of the compared cell group pair}
}
}
\description{
Identifies isoform switching events by comparing every pair of cell groups
using the \code{\link{findMarkers}} function from the scran package and
searching for transcripts of the same gene showing statistically significant
differences in opposite directions.
}
