% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_last_grange.R
\name{get_first_last_grange}
\alias{get_first_last_grange}
\title{Extract the first and last features from a GRangesList}
\usage{
get_first_last_grange(feature_grangeslist)
}
\arguments{
\item{feature_grangeslist}{A GRangesList object.}
}
\value{
A data frame containing following columns:
\describe{
\item{feature_id}{character vector of GRangesList names}
\item{first_grange}{character vector of first feature positions}
\item{last_grange}{character vector of last feature positions}
}
}
\description{
Extracts the first and last feature from each element of a GRangesList
object.
}
\keyword{internal}
