% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_to_intron_bed.R
\name{gtf_to_intron_bed}
\alias{gtf_to_intron_bed}
\title{Create an intron BED file from a GTF file}
\usage{
gtf_to_intron_bed(gtf_file, file)
}
\arguments{
\item{gtf_file}{A string containing a GTF file path.}

\item{file}{A string specifying the output file path.}
}
\value{
Nothing is returned.
}
\description{
Creates an intron BED file from a GTF file.
}
