% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ordered_subset.R
\name{is_ordered_subset}
\alias{is_ordered_subset}
\title{Vector subset check}
\usage{
is_ordered_subset(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical scalar indicating if x is an ordered subset of y.
}
\description{
Checks if a vector is an ordered subset of another vector.
}
\keyword{internal}
