% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isosceles-package.R
\docType{package}
\name{Isosceles}
\alias{Isosceles}
\title{Isosceles: Isoform Single-Cell and Long-read Expression Suite}
\description{
Transcript detection and quantification from ONT reads
}
\author{
Tim Sterne-Weiler \href{mailto:sternewt@gene.com}{sternewt@gene.com}

Michal Kabza \href{mailto:michal.kabza@contractors.roche.com}{michal.kabza@contractors.roche.com}
}
