% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sc_neighbors.R
\name{merge_sc_neighbors}
\alias{merge_sc_neighbors}
\title{Merging neighboring cell TCC values in scRNA-Seq data}
\usage{
merge_sc_neighbors(se_tcc, pca_mat, k = 10, use_annoy = FALSE, ncpu = 1)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{prepare_tcc_se}} function}

\item{pca_mat}{A matrix containing PCA coordinates of each cell}

\item{k}{An integer scalar specifying the number of nearest neighbors to use}

\item{use_annoy}{A logical scalar indicating whether to use the Annoy
algorithm for approximate nearest neighbor identification (recommended for
big datasets)}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization}
}
\value{
A SummarizedExperiment object containing merged TCC data
}
\description{
Prepares a TCC SummarizedExperiment object where count values from the
nearest k neighbors are added to the count values of each cell
}
