% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood_tcc.R
\name{neighborhood_tcc}
\alias{neighborhood_tcc}
\title{Merge the neighboring cell TCC values in scRNA-Seq data}
\usage{
neighborhood_tcc(se_tcc, pca_mat, k = 10, use_annoy = FALSE, ncpu = 1)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{bam_to_tcc}} function.}

\item{pca_mat}{A matrix containing PCA coordinates of each cell.}

\item{k}{An integer scalar specifying the number of nearest neighbors to use.}

\item{use_annoy}{A logical scalar indicating whether to use the Annoy
algorithm for approximate nearest neighbor identification (recommended for
big datasets).}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A SummarizedExperiment object containing merged TCC data.
}
\description{
Prepares a TCC SummarizedExperiment object where count values from the
nearest k neighbors are added to the count values of each cell.
}
