% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_splicing_support_levels.R
\name{plot_splicing_support_levels}
\alias{plot_splicing_support_levels}
\title{Plot spliced read counts by splicing support level}
\usage{
plot_splicing_support_levels(transcript_data)
}
\arguments{
\item{transcript_data}{A named list containing transcript data returned by
the \code{\link{prepare_transcripts}} function.}
}
\value{
A plot object.
}
\description{
Creates a plot showing counts of spliced reads (extracted from the BAM files
using the \code{\link{bam_to_read_structures}} function and processed by the
\code{\link{prepare_transcripts}} function) by their splicing support level.
}
