% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gene_se.R
\name{prepare_gene_se}
\alias{prepare_gene_se}
\title{Prepare a gene-level SummarizedExperiment object}
\usage{
prepare_gene_se(se_tcc)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by a function from
the \code{\link{Isosceles-package}}}
}
\value{
A SummarizedExperiment object containing gene annotation and
quantification data
}
\description{
Prepares a gene-level SummarizedExperiment from TCC data
}
