% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pseudobulk_se.R
\name{prepare_pseudobulk_se}
\alias{prepare_pseudobulk_se}
\title{Prepare a pseudobulk TCC SummarizedExperiment object}
\usage{
prepare_pseudobulk_se(se_tcc, cell_labels)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{prepare_tcc_se}} function}

\item{cell_labels}{A vector or a factor containing cell labels acting as a
grouping variable}
}
\value{
A pseudobulk SummarizedExperiment object containing TCC annotation
and quantification data
}
\description{
Prepares a pseudobulk TCC SummarizedExperiment from TCC data and given cell
labels
}
