% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_psi_se.R
\name{prepare_psi_se}
\alias{prepare_psi_se}
\title{Prepare a PSI SummarizedExperiment object}
\usage{
prepare_psi_se(se)
}
\arguments{
\item{se}{A transcript-level SummarizedExperiment object returned by the
\code{\link{prepare_transcript_se}} function}
}
\value{
A SummarizedExperiment object containing PSI annotation and
quantification data
}
\description{
Prepares a PSI (Percent Spliced In) SummarizedExperiment object
for the given transcript-level SummarizedExperiment object
}
