% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference_annotations.R
\name{prepare_reference_annotations}
\alias{prepare_reference_annotations}
\title{GTF reference annotation data processing}
\usage{
prepare_reference_annotations(gtf_file, is_technical = FALSE)
}
\arguments{
\item{gtf_file}{A string containing a GTF file path}

\item{is_technical}{A boolean scalar specifying if the GTF file describes
technical sequences (e.g. SIRV or ERCC) rather than originating from
Ensembl / GENCODE}
}
\value{
A named list containing following elements:
\describe{
\item{gene_df}{a data frame storing annotated gene data}
\item{transcript_df}{a data frame storing annotated transcript data}
\item{splicing_df}{a data frame storing non-redundant annotated transcript intron structure data}
\item{intron_df}{a data frame storing annotated intron data}
\item{transcript_first_last_df}{a data frame storing confirmed transcript first/last exon/intron positions}
}
}
\description{
Extract required reference annotation from a GTF file
}
\keyword{internal}
