% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference_spliced_transcripts.R
\name{prepare_reference_spliced_transcripts}
\alias{prepare_reference_spliced_transcripts}
\title{Prepare reference spliced transcript data}
\usage{
prepare_reference_spliced_transcripts(
  anno_data,
  bin_size = 50,
  use_full_hash = FALSE
)
}
\arguments{
\item{anno_data}{A list containing genome annotation data returned by
the \code{\link{prepare_reference_annotations}} function.}

\item{bin_size}{An integer scalar specifying the bin size for transcript
start and end position binning.}

\item{use_full_hash}{A logical scalar specifying if full value of the MD5
hash (32 characters) should be used for the stable hash identifier rather
than its 16-character substring. This option should not be used unless you
encounter a hashing collision error (extremely unlikely).}
}
\value{
A named list containing following elements:
\describe{
\item{tx_df}{a data frame storing reference spliced transcript data}
\item{tx_granges}{a GRanges object storing genomic positions of reference spliced transcript}
\item{tx_exon_granges_list}{a GRangesList object storing exon genomic positions of reference spliced transcript}
\item{tx_intron_granges_list}{a GRangesList object storing intron genomic positions of reference spliced transcript}
}
}
\description{
Prepares reference spliced transcript data for further processing.
}
\keyword{internal}
