% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reference_unspliced_transcripts.R
\name{prepare_reference_unspliced_transcripts}
\alias{prepare_reference_unspliced_transcripts}
\title{Prepare reference unspliced transcript data}
\usage{
prepare_reference_unspliced_transcripts(anno_data, bin_size = 50)
}
\arguments{
\item{anno_data}{A list containing genome annotation data returned by
the \code{\link{prepare_reference_annotations}} function.}

\item{bin_size}{An integer scalar specifying the bin size for transcript
start and end position binning.}
}
\value{
A named list containing following elements:
\describe{
\item{tx_df}{a data frame storing reference unspliced transcript data}
\item{tx_granges}{a GRanges object storing genomic positions of reference unspliced transcript}
\item{tx_exon_granges_list}{a GRangesList object storing exon genomic positions of reference unspliced transcript}
\item{tx_intron_granges_list}{a GRangesList object storing intron genomic positions of reference unspliced transcript}
}
}
\description{
Prepares reference unspliced transcript data for further processing.
}
\keyword{internal}
