% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_tcc.R
\name{pseudobulk_tcc}
\alias{pseudobulk_tcc}
\title{Prepare a pseudobulk TCC SummarizedExperiment object}
\usage{
pseudobulk_tcc(se_tcc, cell_labels)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{bam_to_tcc}} function.}

\item{cell_labels}{A vector or a factor containing cell labels acting as a
grouping variable.}
}
\value{
A pseudobulk SummarizedExperiment object containing TCC annotation
and quantification data.
}
\description{
Prepares a pseudobulk TCC SummarizedExperiment from TCC data and given cell
labels.
}
