% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime_tcc.R
\name{pseudotime_tcc}
\alias{pseudotime_tcc}
\title{Merge TCC values using moving window over pseudotime}
\usage{
pseudotime_tcc(
  se_tcc,
  pseudotime,
  trim = 0,
  window_size = 30,
  window_step = 15
)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{prepare_tcc_se}} function.}

\item{pseudotime}{A numeric vector containing the pseudotime values for
each cell. Cells not belonging to the analyzed trajectory should be denoted
using NA values.}

\item{trim}{A numeric scalar specifying the fraction (0 to 0.5) of cells
to be trimmed from each end of the pseudotime spectrum.}

\item{window_size}{An integer scalar specifying the window size.}

\item{window_step}{An integer scalar specifying the window step.}
}
\value{
A SummarizedExperiment object containing TCC data for pseudotime
windows.
}
\description{
Prepares a pseudotime window TCC SummarizedExperiment from TCC data and
pseudotime values.
}
