% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_to_dexseq.R
\name{psi_to_dexseq}
\alias{psi_to_dexseq}
\title{Prepare a PSI count DEXSeqDataSet object}
\usage{
psi_to_dexseq(
  se_psi,
  condition,
  design = ~sample + exon + condition:exon,
  psi_events = NULL,
  remove_redundant_psi = TRUE
)
}
\arguments{
\item{se_psi}{A PSI SummarizedExperiment object returned by the
\code{\link{add_psi_counts}} function.}

\item{condition}{A vector or a factor containing the condition variable
(categorical or continuous) used in the design formula. Alternatively,
a data frame containing multiple variables in separate columns can be used,
in which case the design formula needs to be adjusted by the user.}

\item{design}{A formula which specifies the design of the experiment. See
the DEXSeq package documentation for more information.}

\item{psi_events}{A character vector specifying the PSI events to restrict
the analysis to (ignored if set to NULL).}

\item{remove_redundant_psi}{A logical scalar specifying if PSI events with
redundant count profiles should be removed from the analysis.}
}
\value{
A DEXSeqDataSet object containing PSI count data, suitable for
further analysis using the DEXSeq package.
}
\description{
Prepares a PSI count DEXSeqDataSet object suitable for the analysis of PSI
count changes along the given variable(s) (categorical or continuous).
}
