% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_to_dexseq.R
\name{tcc_to_dexseq}
\alias{tcc_to_dexseq}
\title{Prepare a PSI count DEXSeqDataSet object}
\usage{
tcc_to_dexseq(
  se_tcc,
  pseudotime,
  psi_events = NULL,
  trim = 0,
  window_size = 30,
  window_step = 15,
  remove_redundant_psi = TRUE,
  scale_pseudotime = TRUE,
  ncpu = 1
)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by the
\code{\link{bam_to_tcc}} function.}

\item{pseudotime}{A numeric vector containing the pseudotime values for
each cell. Cells not belonging to the analyzed trajectory should be denoted
using NA values.}

\item{psi_events}{A character vector specifying the PSI events to restrict
the analysis to (ignored if set to NULL).}

\item{trim}{A numeric scalar specifying the fraction (0 to 0.5) of cells
to be trimmed from each end of the pseudotime spectrum.}

\item{window_size}{An integer scalar specifying the window size.}

\item{window_step}{An integer scalar specifying the window step.}

\item{remove_redundant_psi}{A logical scalar specifying if PSI events with
redundant count profiles should be removed from the analysis.}

\item{scale_pseudotime}{A logical scalar specifying if pseudotime values for
the windows should be scaled.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A DEXSeqDataSet object containing PSI count data for pseudotime
windows, suitabe for further analysis using the DEXSeq package.
}
\description{
Aggregates TCC values using pseudotime windows and creates a DEXSeqDataSet
object suitable for the analysis of PSI count changes along given pseudotime
trajectory.
}
