% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_to_gene.R
\name{tcc_to_gene}
\alias{tcc_to_gene}
\title{Prepare a gene-level SummarizedExperiment object}
\usage{
tcc_to_gene(se_tcc)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by a function from
the \code{\link{Isosceles-package}}.}
}
\value{
A SummarizedExperiment object containing gene annotation and
quantification data.
}
\description{
Prepares a gene-level SummarizedExperiment from TCC data.
}
