% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcc_to_transcript.R
\name{tcc_to_transcript}
\alias{tcc_to_transcript}
\title{Prepare a transcript-level SummarizedExperiment object}
\usage{
tcc_to_transcript(
  se_tcc,
  em.maxiter = 250,
  em.conv = 0.01,
  use_length_normalization = TRUE,
  ncpu = 1
)
}
\arguments{
\item{se_tcc}{A TCC SummarizedExperiment object returned by a function from
the \code{\link{Isosceles-package}}.}

\item{em.maxiter}{An integer scalar specifying the maximum number of EM
iterations.}

\item{em.conv}{A numeric scalar specifying the EM convergence threshold.}

\item{use_length_normalization}{A logical scalar specifying if normalization
using effective transcript lengths should be used during EM.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A SummarizedExperiment object containing transcript annotation and
quantification data.
}
\description{
Prepares a transcript-level SummarizedExperiment from TCC data using the EM
algorithm.
}
