% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_to_psi.R
\name{transcript_to_psi}
\alias{transcript_to_psi}
\title{Prepare a PSI SummarizedExperiment object}
\usage{
transcript_to_psi(se, ncpu = 1)
}
\arguments{
\item{se}{A transcript-level SummarizedExperiment object returned by the
\code{\link{tcc_to_transcript}} function.}

\item{ncpu}{An integer scalar specifying the number of cores to use for
multicore parallelization.}
}
\value{
A SummarizedExperiment object containing PSI annotation and
quantification data.
}
\description{
Prepares a PSI (Percent Spliced In) SummarizedExperiment object
for the given transcript-level SummarizedExperiment object. PSI values
are calculated for the following types of regions:
\itemize{
\item{\strong{TSS} - transcription start sites}
\item{\strong{TES} - transcription end sites}
\item{\strong{CE} - core exonic regions}
\item{\strong{RI} - retained intronic regions}
\item{\strong{A5} - 5' alternative exonic regions}
\item{\strong{A3} - 3' alternative exonic regions}
}
TSS and TES positions are calculated based on transcripts' binned start and
end coordinates extracted from their identifiers.
}
