% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_transcripts.R
\name{prepare_transcripts}
\alias{prepare_transcripts}
\title{Transcript data preparation}
\usage{
prepare_transcripts(
  gtf_file,
  genome_fasta_file,
  bam_parsed,
  min_intron_length = 30,
  known_intron_motifs = c("GT-AG"),
  rescue_annotated_introns = FALSE,
  known_intron_granges = NULL,
  min_bam_splice_read_count = 2,
  min_bam_splice_fraction = 0.1,
  bin_size = 50
)
}
\arguments{
\item{gtf_file}{A string containing a GTF file path}

\item{genome_fasta_file}{A string containing a genome FASTA file path}

\item{bam_parsed}{A data frame containing non-redundant read structure data
returned by the \code{\link{extract_read_structures}} function. If NULL,
only reference transcripts are used}

\item{min_intron_length}{An integer scalar specifying the minimal length
of introns to assign strand to}

\item{known_intron_motifs}{A character vector specifying the known intron
motifs}

\item{rescue_annotated_introns}{A logical scalar specifying if introns
found in genome annotations should be kept even if they don't have known
intron motifs}

\item{known_intron_granges}{A GRanges object storing known intron positions
(e.g. from short read data) used for transcript classification. If set to
NULL, only introns from reference annotations are used}

\item{min_bam_splice_read_count}{An integer scalar specifying the read count
threshold for splice sites confirmed by aligned reads}

\item{min_bam_splice_fraction}{A numeric scalar specifying the minimum
connectivity fraction to a known splice site for splice sites confirmed by
aligned reads}

\item{bin_size}{An integer scalar specifying the bin size for transcript
start and end position binning}
}
\value{
A named list containing following elements:
\describe{
\item{tx_df}{a data frame storing extracted transcript data}
\item{tx_granges}{a GRanges object storing genomic positions of extracted transcript}
\item{tx_exon_granges_list}{a GRangesList object storing exon genomic positions of extracted transcript}
\item{tx_intron_granges_list}{a GRangesList object storing intron genomic positions of extracted transcript}
}
}
\description{
Prepare transcript data (reference and extracted from the BAM files) for
further analysis
}
